+++ predico/services/resource/base_resource.py
@@ -6,3 +6,17 @@ from dataclasses import dataclass
 class Resource:
     rtype: str
     id: str
+
+
+class Resources(dict):
+    """
+
+    Custom dictionary container for storing resources.
+
+    This was first created because we needed a marker to make
+    Resources an injectable. Later we might do some things like
+    caching.
+
+    """
+
+    pass