+++ tests/integration/resource/test_resource.py
@@ -1,3 +1,6 @@
+from predico.services.request.common_request import CommonRequest
+
+
 def test_resources_empty(rs):
     assert {} == rs.resources
 
@@ -21,3 +24,15 @@ def test_request_resource(services, test_resources):
 
     # Request: Did the request get the correct one?
     assert 'more/contact' == request.resource.id
+
+
+def test_resources_adapter(services, test_resources):
+    request_service = services['request']
+    request: CommonRequest = request_service.make_request(
+        resourceid='more/contact'
+    )
+
+    resource_service = services['resource']
+    resources = resource_service.resources
+    resource = resources['more/contact']
+    assert request.resource.id == resource.id